requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
multidict==6.0.5
ngrok==1.2.0
packaging==24.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
twilio==9.0.5
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
